move test code out
